% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini2.R
\name{gini2}
\alias{gini2}
\title{Compute the Gini Coefficient}
\usage{
gini2(x, weights = rep(1, length = length(x)))
}
\arguments{
\item{x}{a vector containing at least non-negative elements}

\item{weights}{an optional vector of sample weights for x}
}
\value{
The Gini coefficient (between 0 and 1).
}
\description{
Computes the Gini coefficient based on (possibly weighted) sample data.
}
\examples{
# generate vector (of incomes)
x <- c(541, 1463, 2445, 3438, 4437, 5401, 6392, 8304, 11904, 22261)
# compute Gini coefficient
gini2(x)
# generate a vector of weights.
w <- runif(n=length(x))
gini2(x,w)
}
