% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kakwani.R
\name{kakwani}
\alias{kakwani}
\title{Title}
\usage{
kakwani(var1, var2, weights = 1)
}
\arguments{
\item{var1}{Var1}

\item{var2}{Var2}

\item{weights}{Weights}
}
\value{
Value
}
\description{
Title
}
\examples{
This is an example
}
