% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Round up to the nearest unit}
\usage{
round2(x, digits = 0)
}
\arguments{
\item{x}{A numeric vector to round up.}

\item{digits}{The desired number of decimals.}
}
\value{
A numeric vector.
}
\description{
Round up to the nearest unit
}
\examples{
x <- c(0.5,1.5,2.5)
round2(x, digits = 0)
}
