% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kakwani.R
\name{kakwani}
\alias{kakwani}
\title{Compute the Kakwani Coefficient}
\usage{
kakwani(var1, var2, weights = 1)
}
\arguments{
\item{var1}{a numeric vector}

\item{var2}{a numeric vector}

\item{weights}{an optional vector of sample weights}
}
\value{
the Kakwani coefficient.
}
\description{
Compute the Kakwani Coefficient
}
\examples{
# This is an example
# generate vectors (of gross income and post tax income)
var1 <- c(541, 1463, 2445, 3438, 4437, 5401, 6392, 8304, 11904, 22261)
var2 <- c(487, 1284, 2092, 2865, 3599, 4261, 4901, 6182, 8597, 15583)
# compute Kakwani coefficient
kakwani(var1,var2)
# generate a vector of weights.
weights <- runif(n=length(var1))
kakwani(var1,var2,weights)
}
