% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Rounding of Numbers Away From Zero}
\usage{
round2(x, digits = 0)
}
\arguments{
\item{x}{a numeric vector to round.}

\item{digits}{the desired number of decimals, by default 0.}
}
\value{
A numeric vector.
}
\description{
\code{round2} adopts the convention of rounding away from zero.
}
\examples{
x <- c(-2.5,-1.5,-0.5,0.5,1.5,2.5)
round2(x, digits = 0)
}
